refactor: change all components imports to reference components index file
