+++ LanguageApp/app/components/AuthErrorMsg.js
@@ -3,7 +3,7 @@
 import React from "react";
 import { StyleSheet } from "react-native";
 
-import AppText from "./AppText";
+import { AppText } from "./";
 
 const AuthErrorMsg = ({ error, visible }) => {
   if (!error || !visible) {
