+++ LanguageApp/app/components/ErrorMessage.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { View, StyleSheet } from "react-native";
-import AppText from "./AppText";
+import { AppText } from "./";
 
 function ErrorMessage({ style, message }) {
   return (
