+++ LanguageApp/app/components/FormMessages.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { StyleSheet } from "react-native";
-import AppText from "./AppText";
+import { AppText } from "./";
 
 import colors from "../config/colors";
