+++ LanguageApp/app/components/ScenarioImage.js
@@ -1,8 +1,7 @@
 import React, { useState, useEffect } from "react";
 import { Image, StyleSheet, View } from "react-native";
 import { storage } from "../../firebaseSetup";
-import LoadingSign from "./LoadingSign";
-import ErrorMessage from "./ErrorMessage";
+import { LoadingSign, ErrorMessage } from "./";
 
 function ScenarioImage({ uriLink }) {
   const [isLoading, setIsLoading] = useState();
