+++ LanguageApp/app/screens/UserTypeScreen.js
@@ -1,8 +1,6 @@
 import { StyleSheet, View } from "react-native";
 import React, { useContext } from "react";
-
-import Screen from "../components/Screen";
-import AppButton from "../components/AppButton";
+import { Screen, AppButton } from "../components";
 import routes from "../navigation/routes";
 import colors from "../config/colors";
